Support setters for traits not explicitly used in a class.
If a trait uses another trait, it will not show up in the return array
from class_uses() when called on a class that uses only that first
class.

Fortunately, we can call class_uses() on classes and traits. Therefore,
it is somewhat trivial to recursively call class_uses() on all traits
used in a class, to get the true absolute list of all traits used.