Add a new LazyCallable type for injecting callable services
It's possible for code to expect to take a callable and invoke it
multiple times (sequentially, not recursively) with parameters that are
only known at calltime. This patch makes it possible to use services
configured in the dependency injection container as callables that fit
the above definition.

The LazyCallable type differs from the Lazy type in two ways:

1. Lazy expects all callable parameters to be configured when it is
   created. LazyCallable expects all callable parameters to be supplied
   at calltime.
2. Lazy expects to only be called once. LazyCallable expects to be
   called multiple times.